Move sidebar logic to theme
